Extracting duplicate methods between SiteSettingsFragment and PublicizeManageConnectionsFragment to abstract class SettingsFragment
